Login refactor to address routing bug (#57)
* chore: login refactor for routing
