+++ src/hooks.tsx
@@ -1,19 +1,11 @@
-import { useRouter } from 'next/router';
-import React, { useState, useEffect } from 'react';
+import { useRouter } from 'next/navigation';
+import { useState, useEffect } from 'react';
 import { getSettings } from '@/api/users/settings';
-import { getFood } from './api/food';
-import { RecipeIngredient } from './app/recipe/page';
 import { SettingsContextInterface, defaultSettings } from './context';
-import { DefinedFoodObject } from './interfaces/FoodObject';
-
-export const logout = () => {
-  const router = useRouter();
-  localStorage.removeItem('uid');
-  localStorage.removeItem('email');
-  router.push('/');
-};
 
 export const useLogin = () => {
+  const router = useRouter();
+
   const [uid, setUid] = useState<string>('');
   const [email, setEmail] = useState<string>('');
 
@@ -22,7 +14,9 @@ export const useLogin = () => {
       setUid(localStorage.getItem('uid') || 'ERROR');
       setEmail(localStorage.getItem('email') || 'ERROR');
     } else {
-      logout();
+      localStorage.removeItem('uid');
+      localStorage.removeItem('email');
+      router.push('/');
     }
   }, []);
 
@@ -33,6 +27,8 @@ export const useLogin = () => {
 };
 
 export const useSettings = async (uid: string) => {
+  const router = useRouter();
+
   const [settings, setUserSettings] = useState<SettingsContextInterface>(defaultSettings);
 
   async function fetchSettings() {
@@ -43,7 +39,9 @@ export const useSettings = async (uid: string) => {
 
   useEffect(() => {
     if (uid == 'ERROR') {
-      logout();
+      localStorage.removeItem('uid');
+      localStorage.removeItem('email');
+      router.push('/');
     } else {
       if (uid != '') {
         fetchSettings();
